+++ worker.js
@@ -390,7 +390,362 @@ EventEmitter.prototype.listeners = function(type) {
 	return this._events[type];
 };
 
-local.util.EventEmitter = EventEmitter;// http://jsperf.com/cloning-an-object/2
+local.util.EventEmitter = EventEmitter;// Helpers
+// =======
+
+if (typeof CustomEvent === 'undefined') {
+	// CustomEvent shim (safari)
+	// thanks to netoneko https://github.com/maker/ratchet/issues/101
+	CustomEvent = function(type, eventInitDict) {
+		var event = document.createEvent('CustomEvent');
+
+		event.initCustomEvent(type, eventInitDict['bubbles'], eventInitDict['cancelable'], eventInitDict['detail']);
+		return event;
+	};
+}
+
+// EXPORTED
+// searches up the node tree for an element
+function findParentNode(node, test) {
+	while (node) {
+		if (test(node)) { return node; }
+		node = node.parentNode;
+	}
+	return null;
+}
+
+findParentNode.byTag = function(node, tagName) {
+	return findParentNode(node, function(elem) {
+		return elem.tagName == tagName;
+	});
+};
+
+findParentNode.byClass = function(node, className) {
+	return findParentNode(node, function(elem) {
+		return elem.classList && elem.classList.contains(className);
+	});
+};
+
+findParentNode.byElement = function(node, element) {
+	return findParentNode(node, function(elem) {
+		return elem === element;
+	});
+};
+
+findParentNode.thatisFormRelated = function(node) {
+	return findParentNode(node, function(elem) {
+		return !!elem.form;
+	});
+};
+
+// combines parameters as objects
+// - precedence is rightmost
+//     reduceObjects({a:1}, {a:2}, {a:3}) => {a:3}
+function reduceObjects() {
+	var objs = Array.prototype.slice.call(arguments);
+	var acc = {}, obj;
+	while (objs.length) {
+		obj = objs.shift();
+		if (!obj) { continue; }
+		for (var k in obj) {
+			if (typeof obj[k] == 'undefined' || obj[k] === null) { continue; }
+			if (typeof obj[k] == 'object' && !Array.isArray(obj[k])) {
+				acc[k] = reduceObjects(acc[k], obj[k]);
+			} else {
+				acc[k] = obj[k];
+			}
+		}
+	}
+	return acc;
+}
+
+// EXPORTED
+// dispatches a request event, stopping the given event
+function dispatchRequestEvent(targetElem, request) {
+	var re = new CustomEvent('request', { bubbles:true, cancelable:true, detail:request });
+	targetElem.dispatchEvent(re);
+}
+
+// EXPORTED
+// submit helper, makes it possible to find the button which triggered the submit
+function trackFormSubmitter(node) {
+	var elem = findParentNode.thatisFormRelated(node);
+	if (elem) {
+		for (var i=0; i < elem.form.length; i++) {
+			elem.form[i].setAttribute('submitter', null);
+		}
+		elem.setAttribute('submitter', '1');
+	}
+}
+
+// EXPORTED
+// extracts request from any given element
+function extractRequest(targetElem, containerElem) {
+	var requests = { form:{}, fieldset:{}, elem:{} };
+	var fieldset = null, form = null;
+
+	// find parent fieldset
+	if (targetElem.tagName === 'FIELDSET') {
+		fieldset = targetElem;
+	} else if (targetElem.tagName !== 'FORM') {
+		fieldset = findParentNode.byTag(targetElem, 'FIELDSET');
+	}
+
+	// find parent form
+	if (targetElem.tagName === 'FORM') {
+		form = targetElem;
+	} else {
+		// :TODO: targetElem.form may be a simpler alternative
+		var formId = targetElem.getAttribute('form') || (fieldset ? fieldset.getAttribute('form') : null);
+		if (formId) {
+			form = containerElem.querySelector('#'+formId);
+		}
+		if (!form) {
+			form = findParentNode.byTag(targetElem, 'FORM');
+		}
+	}
+
+	// extract payload
+	var payload = extractRequestPayload(targetElem, form);
+
+	// extract form headers
+	if (form) {
+		requests.form = extractRequest.fromForm(form, targetElem);
+	}
+
+	// extract fieldset headers
+	if (fieldset) {
+		requests.fieldset = extractRequest.fromFormElement(fieldset);
+	}
+
+	// extract element headers
+	if (targetElem.tagName === 'A') {
+		requests.elem = extractRequest.fromAnchor(targetElem);
+	} else if (['FORM','FIELDSET'].indexOf(targetElem.tagName) === -1) {
+		requests.elem = extractRequest.fromFormElement(targetElem);
+	}
+
+	// combine then all, with precedence given to rightmost objects in param list
+	var req = reduceObjects(requests.form, requests.fieldset, requests.elem);
+	var payloadWrapper = {};
+	payloadWrapper[/GET/i.test(req.method) ? 'query' : 'body'] = payload;
+	return reduceObjects(req, payloadWrapper);
+}
+
+// EXPORTED
+// extracts request parameters from an anchor tag
+extractRequest.fromAnchor = function(node) {
+
+	// get the anchor
+	node = findParentNode.byTag(node, 'A');
+	if (!node || !node.attributes.href || node.attributes.href.value.charAt(0) == '#') { return null; }
+
+	// pull out params
+	var request = {
+		// method  : 'get',
+		url     : node.attributes.href.value,
+		target  : node.getAttribute('target'),
+		headers : { accept:node.getAttribute('type') }
+	};
+	return request;
+};
+
+// EXPORTED
+// extracts request parameters from a form element (inputs, textareas, etc)
+extractRequest.fromFormElement = function(node) {
+	// :TODO: search parent for the form-related element?
+	//        might obviate the need for submitter-tracking
+
+	// pull out params
+	var request = {
+		method  : node.getAttribute('formmethod'),
+		url     : node.getAttribute('formaction'),
+		target  : node.getAttribute('formtarget'),
+		headers : {
+			'content-type' : node.getAttribute('formenctype'),
+			accept         : node.getAttribute('formaccept')
+		}
+	};
+	return request;
+};
+
+// EXPORTED
+// extracts request parameters from a form
+extractRequest.fromForm = function(form, submittingElem) {
+
+	// find the submitter, if the submitting element is not form-related
+	if (submittingElem && !submittingElem.form) {
+		for (var i=0; i < form.length; i++) {
+			var elem = form[i];
+			if (elem.getAttribute('submitter') == '1') {
+				submittingElem = elem;
+				elem.setAttribute('submitter', '0');
+				break;
+			}
+		}
+	}
+
+	var requests = { submitter:{}, form:{} };
+	// extract submitting element headers
+	if (submittingElem) {
+		requests.submitter = {
+			method  : submittingElem.getAttribute('formmethod'),
+			url     : submittingElem.getAttribute('formaction'),
+			target  : submittingElem.getAttribute('formtarget'),
+			headers : {
+				'content-type' : submittingElem.getAttribute('formenctype'),
+				accept         : submittingElem.getAttribute('formaccept')
+			}
+		};
+	}
+	// extract form headers
+	requests.form = {
+		method  : form.getAttribute('method'),
+		url     : form.getAttribute('action'),
+		target  : form.getAttribute('target'),
+		headers : {
+			'content-type' : form.getAttribute('enctype') || form.enctype,
+			'accept'       : form.getAttribute('accept')
+		}
+	};
+	if (form.acceptCharset) { requests.form.headers.accept = form.acceptCharset; }
+
+	// combine, with precedence to the submitting element
+	var request = reduceObjects(requests.form, requests.submitter);
+
+	// strip the base URI
+	// :TODO: needed?
+	/*var base_uri = window.location.href.split('#')[0];
+	if (target_uri.indexOf(base_uri) != -1) {
+		target_uri = target_uri.substring(base_uri.length);
+		if (target_uri.charAt(0) != '/') { target_uri = '/' + target_uri; }
+	}*/
+
+	return request;
+};
+
+// EXPORTED
+// serializes all form elements beneath and including the given element
+// - `targetElem`: container element, will reject the field if not within (optional)
+// - `form`: an array of HTMLElements or a form field (they behave the same for iteration)
+// - `opts.nofiles`: dont try to read files in file fields? (optional)
+function extractRequestPayload(targetElem, form, opts) {
+	if (!opts) opts = {};
+
+	// iterate form elements
+	var data = {};
+	if (!opts.nofiles)
+		data.__fileReads = []; // an array of promises to read <input type=file>s
+	for (var i=0; i < form.length; i++) {
+		var elem = form[i];
+
+		// skip if it doesnt have a name
+		if (!elem.name)
+			continue;
+
+		// skip if not a child of the target element
+		if (targetElem && !findParentNode.byElement(elem, targetElem))
+			continue;
+
+		// pull value if it has one
+		var isSubmittingElem = elem.getAttribute('submitter') == '1';
+		if (elem.tagName === 'BUTTON') {
+			if (isSubmittingElem) {
+				// don't pull from buttons unless recently clicked
+				data[elem.name] = elem.value;
+			}
+		} else if (elem.tagName === 'INPUT') {
+			switch (elem.type.toLowerCase()) {
+				case 'button':
+				case 'submit':
+					if (isSubmittingElem) {
+						// don't pull from buttons unless recently clicked
+						data[elem.name] = elem.value;
+					}
+					break;
+				case 'checkbox':
+					if (elem.checked) {
+						// don't pull from checkboxes unless checked
+						data[elem.name] = (data[elem.name] || []).concat(elem.value);
+					}
+					break;
+				case 'radio':
+					if (elem.getAttribute('checked') !== null) {
+						// don't pull from radios unless selected
+						data[elem.name] = elem.value;
+					}
+					break;
+				case 'file':
+					// read the files
+					if (opts.nofiles)
+						break;
+					if (elem.multiple) {
+						for (var i=0, f; f = elem.files[i]; i++)
+							readFile(data, elem, elem.files[i], i);
+						data[elem.name] = [];
+						data[elem.name].length = i;
+					} else {
+						readFile(data, elem, elem.files[0]);
+					}
+					break;
+				default:
+					data[elem.name] = elem.value;
+					break;
+			}
+		} else
+			data[elem.name] = elem.value;
+	}
+
+	return data;
+}
+
+// INTERNAL
+// file read helpers
+function readFile(data, elem, file, index) {
+	if (!file) return; // no value set
+	var reader = new FileReader();
+	reader.onloadend = readFileLoadEnd(data, elem, file, index);
+	reader.readAsDataURL(file);
+}
+function readFileLoadEnd(data, elem, file, index) {
+	// ^ this avoids a closure circular reference
+	var promise = local.promise();
+	data.__fileReads.push(promise);
+	return function(e) {
+		var obj = {
+			content: e.target.result || null,
+			name: file.name,
+			formattr: elem.name,
+			size: file.size,
+			type: file.type,
+			lastModifiedDate: file.lastModifiedDate
+		};
+		if (typeof index != 'undefined')
+			obj.formindex = index;
+		promise.fulfill(obj);
+	};
+}
+function finishPayloadFileReads(request) {
+	var fileReads = (request.body) ? request.body.__fileReads :
+					((request.query) ? request.query.__fileReads : []);
+	return local.promise.bundle(fileReads).then(function(files) {
+		if (request.body) delete request.body.__fileReads;
+		if (request.query) delete request.query.__fileReads;
+		files.forEach(function(file) {
+			if (typeof file.formindex != 'undefined')
+				request.body[file.formattr][file.formindex] = file;
+			else request.body[file.formattr] = file;
+		});
+		return request;
+	});
+}
+
+local.util.findParentNode = findParentNode;
+local.util.trackFormSubmitter = trackFormSubmitter;
+local.util.dispatchRequestEvent = dispatchRequestEvent;
+local.util.extractRequest = extractRequest;
+local.util.extractRequestPayload = extractRequestPayload;
+local.util.finishPayloadFileReads = finishPayloadFileReads;// http://jsperf.com/cloning-an-object/2
 local.util.deepClone = function(obj) {
 	return JSON.parse(JSON.stringify(obj));
 };})();// Local HTTP
@@ -3606,4 +3961,76 @@ if (typeof this.local == 'undefined')
 (function() {// Aliases to common functions
 local.dispatch = local.web.dispatch;
 local.subscribe = local.web.subscribe;
-local.navigator = local.web.navigator;})();
+local.navigator = local.web.navigator;// Standard DOM Events
+// ===================
+
+// bindRequestEvents()
+// ===================
+// EXPORTED
+// Converts 'click' and 'submit' events into custom 'request' events
+// - within the container, all 'click' and 'submit' events will be consumed
+// - 'request' events will be dispatched by the original dispatching element
+// Parameters:
+// - `container` must be a valid DOM element
+// - `options` may disable event listeners by setting `links` or `forms` to false
+function bindRequestEvents(container, options) {
+	container.__localEventHandlers = [];
+	options = options || {};
+
+	var handler;
+	if (options.links !== false) {
+		handler = { name: 'click', handleEvent: Local__clickHandler, container: container };
+		container.addEventListener('click', handler, true);
+		container.__localEventHandlers.push(handler);
+	}
+	if (options.forms !== false) {
+		handler = { name: 'submit', handleEvent: Local__submitHandler, container: container };
+		container.addEventListener('submit', handler, true);
+	}
+}
+
+// unbindRequestEvents()
+// =====================
+// EXPORTED
+// Stops listening to 'click' and 'submit' events
+function unbindRequestEvents(container) {
+	if (container.__localEventHandlers) {
+		container.__localEventHandlers.forEach(function(handler) {
+			container.removeEventListener(handler.name, handler);
+		});
+		delete container.__localEventHandlers;
+	}
+}
+
+// INTERNAL
+// transforms click events into request events
+function Local__clickHandler(e) {
+	if (e.button !== 0) { return; } // handle left-click only
+	local.util.trackFormSubmitter(e.target);
+	var request = local.util.extractRequest.fromAnchor(e.target);
+	if (request && ['_top','_blank'].indexOf(request.target) !== -1) { return; }
+	if (request) {
+		e.preventDefault();
+		e.stopPropagation();
+		local.util.dispatchRequestEvent(e.target, request);
+		return false;
+	}
+}
+
+// INTERNAL
+// transforms submit events into request events
+function Local__submitHandler(e) {
+	var request = local.util.extractRequest(e.target, this.container);
+	if (request && ['_top','_blank'].indexOf(request.target) !== -1) { return; }
+	if (request) {
+		e.preventDefault();
+		e.stopPropagation();
+		local.util.finishPayloadFileReads(request).then(function() {
+			local.util.dispatchRequestEvent(e.target, request);
+		});
+		return false;
+	}
+}
+
+local.bindRequestEvents = bindRequestEvents;
+local.unbindRequestEvents = unbindRequestEvents;})();